# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "WaveOne API",
        "contact": {
            "name": "Aneep Tandel",
            "url": "https://github.com/aneepct"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/register.{format?}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns registration response.",
                "operationId": "api.register.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "description": "Phone Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/auth/v2/register.{format?}": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Returns registration v2 response.",
                "operationId": "api.register.index",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "name": "format?",
                        "in": "path",
                        "description": "Response format.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "description": "Phone Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Country overview."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
...
}